chore: update 11 crates to Rust 2024 edition (#19258)
## Which issue does this PR close?

This addresses part of https://github.com/apache/datafusion/issues/15804
but does not close it.

## Rationale for this change

Now that we are on MSRV 1.88 we can use rust edition 2024, which brings
let chains and other nice features. It also improves `unsafe` checking.
In order to introduce these changes in slower way instead of one massive
PR that is too difficult to manage we are updating a few crates at a
time.

## What changes are included in this PR?

Updates these crates to 2024.

- datafusion-catalog
- datafusion-datasource-arrow
- datafusion-datasource-avro
- datafusion-datasource-csv
- datafusion-datasource-json
- datafusion-datasource-parquet
- datafusion-pruning
- datafusion-functions-table
- datafusion-physical-optimizer
- datafusion-spark
- datafusion-catalog-listing

## Are these changes tested?

Existing unit tests. There are no functional code changes.

## Are there any user-facing changes?

None.

## Note

It is recommended to review with the ignore whitespace setting:
https://github.com/apache/datafusion/pull/19258/files?w=1

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>